Moving filter form to more precise location
The filter form is something that I believe is only available when an
authenticated patron visits the work areas page.
